Removed redundant optimization to sendPlayerCountUpdate
The bot manager automatically updates the player counts when a bot spawns. If a bot was to be revived during a match using one of the compatable revive mods, the player count would be marked to update on the server and sent to all clients. For some reason this update was set to only happen every half a second, which could be really bad since a player could die in that time and would result in the game thinking all players are dead when they are clearly not.