Fixed a logic error in findSafePathToTarget
There was a logic error where the bot was checking the player position rather than the node position when checking for a safe node to take cover at.

This only affected bots that were checking for cover when they were caught out in the open and every available path was considered dangerous.